Refactor. Add --dummy option.
